// static/scripts/server-types.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type EntryReceipt = { timeStamp: string, id: string, };

export type Guestbook = { guestbook: Array<GuestbookEntry>, };

export type GuestbookEntry = { id?: string, timeStamp?: string, name: string, note: string, };

export type ListInfo = { listName: string, authorUser: string, attrs: Array<string>, };

export type ListRow = { totalRows: number, rowData: string, };

export type WebpageHit = { timeStamp: string, userAgent: string, };